+++ classes/class-gallery.php
@@ -24,7 +24,7 @@ class Mai_Gallery {
 	/**
 	 * Mai_Gallery constructor.
 	 *
-	 * @since TBD
+	 * @since 1.0.0
 	 *
 	 * @param array $args Gallery args.
 	 *
@@ -36,7 +36,9 @@ class Mai_Gallery {
 				'preview'                => false,
 				'align'                  => '',
 				'class'                  => '',
+				'links'                  => false,
 				'images'                 => [],
+				'images_links'           => [],
 				'image_orientation'      => 'landscape',
 				'image_size'             => 'sm',
 				'shadow'                 => false,
@@ -59,7 +61,9 @@ class Mai_Gallery {
 		$args['preview']           = mai_sanitize_bool( $args['preview'] );
 		$args['align']             = esc_attr( $args['align'] );
 		$args['class']             = esc_html( $args['class'] );
+		$args['links']             = mai_sanitize_bool( $args['links'] );
 		$args['images']            = $args['images'] ? array_map( 'absint', (array) $args['images'] ) : [];
+		$args['images_links']      = $this->sanitize_links( $args['images_links'] ); // Later in loop.
 		$args['image_orientation'] = esc_html( $args['image_orientation'] );
 		$args['image_size']        = esc_html( $args['image_size'] );
 		$args['shadow']            = mai_sanitize_bool( $args['shadow'] );
@@ -75,7 +79,7 @@ class Mai_Gallery {
 	/**
 	 * Displays gallery.
 	 *
-	 * @since TBD
+	 * @since 1.0.0
 	 *
 	 * @return void
 	 */
@@ -86,15 +90,19 @@ class Mai_Gallery {
 	/**
 	 * Gets gallery.
 	 *
-	 * @since TBD
+	 * @since 1.0.0
 	 *
 	 * @return string
 	 */
 	function get() {
 		static $count = 1;
-		$html         = '';
 
-		if ( ! $this->args['images'] ) {
+		$html       = '';
+		$images     = [];
+		$has_images = ! $this->args['links'] && $this->args['images'];
+		$has_links  = $this->args['links'] && $this->args['images_links'];
+
+		if ( ! ( $has_images || $has_links ) ) {
 			if ( $this->args['preview'] ) {
 				$text  = __( 'Add gallery images in the block sidebar settings.', 'mai-galleries' );
 				$html .= sprintf( '<p style="display:flex;justify-content:center;align-items:center;color:var(--body-color);font-family:var(--body-font-family);font-weight:var(--body-font-weight);font-size:var(--body-font-size);opacity:0.62;"><span class="dashicons dashicons-format-gallery"></span>&nbsp;&nbsp;%s</p>', $text );
@@ -103,9 +111,30 @@ class Mai_Gallery {
 			return $html;
 		}
 
-		if ( ! $this->args['preview'] && $this->args['lightbox'] ) {
-			$this->enqueue_lightbox_styles();
+		// Add inline CSS.
+		$suffix = $this->get_suffix();
+		$url    = MAI_GALLERIES_PLUGIN_URL . "assets/css/mai-galleries{$suffix}.css";
+		$path   = MAI_GALLERIES_PLUGIN_DIR . "assets/css/mai-galleries{$suffix}.css";
+		wp_enqueue_style( 'mai-galleries', $url );
+		wp_style_add_data( 'mai-galleries', 'path', $path );
+
+		// Maybe enqueue lightbox scripts styles. Too much CSS and no need to load inline since it won't be used until launched.
+		if ( ! $this->args['preview'] && ! $this->args['links'] && $this->args['lightbox'] ) {
+			$lightbox = true;
 			$this->enqueue_lightbox_scripts();
+			$this->enqueue_lightbox_styles();
+		}
+
+		// Build array to match repeater, for easier loop later.
+		if ( $has_images ) {
+			foreach ( $this->args['images'] as $image_id ) {
+				$images[] = [
+					'image' => $image_id,
+					'url'   => '',
+				];
+			}
+		} else {
+			$images = $this->args['images_links'];
 		}
 
 		$atts = [
@@ -147,10 +176,10 @@ class Mai_Gallery {
 			]
 		);
 
-			$html .= $this->get_css();
-
-			foreach ( $this->args['images'] as $image_id ) {
-				$image = wp_get_attachment_image(
+			foreach ( $images as $data ) {
+				$image_id = $data['image'];
+				$url      = $data['url'];
+				$image    = wp_get_attachment_image(
 					$image_id,
 					$this->image_size,
 					false,
@@ -165,9 +194,15 @@ class Mai_Gallery {
 					continue;
 				}
 
+				// Caption.
 				$caption = wp_get_attachment_caption( $image_id );
 
-				if ( $this->args['lightbox'] ) {
+				// Links.
+				if ( $this->args['links'] && $url ) {
+					$image = sprintf( '<a class="mai-gallery-image-link" href="%s">%s</a>', esc_url( $url ), $image );
+				}
+				// Lightbox.
+				elseif ( ! $this->args['links'] && $this->args['lightbox'] ) {
 					$href   = wp_get_attachment_image_url( $image_id, 'medium' );
 					$sizes  = wp_get_attachment_image_sizes( $image_id, 'cover' );
 					$srcset = wp_get_attachment_image_srcset( $image_id, 'cover' );
@@ -181,6 +216,7 @@ class Mai_Gallery {
 					);
 				}
 
+				// Add caption.
 				if ( $caption ) {
 					$image .= sprintf( '<figcaption class="mai-gallery-image-caption">%s</figcaption>', $caption );
 				}
@@ -219,9 +255,32 @@ class Mai_Gallery {
 	}
 
 	/**
+	 * Sanitize image links.
+	 *
+	 * @since 1.2.0
+	 *
+	 * @param array $links
+	 *
+	 * @return array
+	 */
+	function sanitize_links( $links ) {
+		if ( ! $links ) {
+			return $links;
+		}
+
+		foreach ( $links as $index => $data ) {
+			if ( ! isset( $data['image'] ) || ! isset( $data['url'] ) || empty( $data['image'] ) ) {
+				unset( $links[ $index ] );
+			}
+		}
+
+		return $links;
+	}
+
+	/**
 	 * Gets the image size.
 	 *
-	 * @since TBD
+	 * @since 1.0.0
 	 *
 	 * @return string
 	 */
@@ -271,31 +330,6 @@ class Mai_Gallery {
 			[],
 			MAI_GALLERIES_VERSION . '.' . date( 'njYHi', filemtime( MAI_GALLERIES_PLUGIN_DIR . sprintf( 'assets/css/mai-galleries-lightbox%s.css', $this->get_suffix() ) ) )
 		);
-	}
-
-	/**
-	 * Gets toc css link if it hasn't been loaded yet.
-	 *
-	 * @since 0.1.0
-	 *
-	 * @return string
-	 */
-	public function get_css() {
-		static $first = false;
-
-		if ( $first ) {
-			return;
-		}
-
-		$css = file_get_contents( MAI_GALLERIES_PLUGIN_DIR . sprintf( 'assets/css/mai-galleries%s.css', $this->get_suffix() ) );
-
-		if ( $this->args['preview'] ) {
-			$css .= '.mai-gallery a {pointer-events: none;}';
-		}
-
-		$first = true;
-
-		return sprintf( '<style>%s</style>', $css );
 	}
 
 	/**